requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
black==23.7.0
cachetools==5.3.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.4
django-extensions==3.2.3
djangorestframework==3.14.0
executing==1.2.0
filelock==3.12.2
flake8==6.1.0
httpie==3.2.2
idna==3.4
ipython==8.14.0
jedi==0.19.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
pluggy==1.2.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.16.1
pyproject-api==1.5.3
PySocks==1.7.1
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.5.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
tox==4.8.0
traitlets==5.9.0
urllib3==2.0.4
virtualenv==20.24.3
wcwidth==0.2.6
